Added logout button and moved routes into app.js
