+++ client/src/components/404/404.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+
+const BackWoods404 = () => (
+  <div>
+    <h1>404</h1>
+  </div>
+)
+
+export default BackWoods404;
